requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.1
filetype==1.2.0
fonttools==4.61.1
fsspec==2025.12.0
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typing_extensions==4.15.0
ultralytics==8.3.241
ultralytics-thop==2.0.18
urllib3==2.6.2
